+++ R/benchmark_grid.R
@@ -4,12 +4,21 @@
 #' Takes a lists of [Task], a list of [Learner] and a list of [Resampling] to
 #' generate a design in an [expand.grid()] fashion (a.k.a. cross join or Cartesian product).
 #'
-#' Resampling strategies are not allowed to be instantiated when passing the argument, and instead will be instantiated per task internally.
-#' The only exception to this rule applies if all tasks have exactly the same number of rows, and the resamplings are all instantiated for such tasks.
+#' There are two modes of operation, depending on the flag `paired`.
+#'
+#' * With `paired` set to `FALSE` (default), resampling strategies are not allowed to be instantiated, and instead will be instantiated per task internally.
+#'   The only exception to this rule applies if all tasks have exactly the same number of rows, and the resamplings are all instantiated for such tasks.
+#'   The grid will be generated based on the Cartesian product of tasks, learners, and resamplings.
+#' * With `paired` set to `TRUE`, tasks and resamplings are treated as pairs.
+#'   I.e., you must provide as many tasks as corresponding instantiated resamplings.
+#'   The grid will be generated based on the Cartesian product of learners and pairs.
 #'
 #' @param tasks (list of [Task]).
 #' @param learners (list of [Learner]).
 #' @param resamplings (list of [Resampling]).
+#' @param paired (`logical(1)`)\cr
+#'   Set this to `TRUE` if the resamplings are instantiated on the tasks, i.e., the tasks and resamplings are paired.
+#'   You need to provide the same number of tasks and instantiated resamplings.
 #'
 #' @return ([data.table::data.table()]) with the cross product of the input vectors.
 #'
@@ -26,6 +35,17 @@
 #' benchmark(grid)
 #' }
 #'
+#' # paired
+#' learner = lrn("classif.rpart")
+#' task1 = tsk("penguins")
+#' task2 = tsk("german_credit")
+#' res1 = rsmp("holdout")
+#' res2 = rsmp("holdout")
+#' res1$instantiate(task1)
+#' res2$instantiate(task2)
+#' design = benchmark_grid(list(task1, task2), learner, list(res1, res2), paired = TRUE)
+#' print(design)
+#'
 #' # manual construction of the grid with data.table::CJ()
 #' grid = data.table::CJ(task = tasks, learner = learners,
 #'   resampling = resamplings, sorted = FALSE)
@@ -37,33 +57,61 @@
 #' \dontrun{
 #' benchmark(grid)
 #' }
-benchmark_grid = function(tasks, learners, resamplings) {
+#'
+benchmark_grid = function(tasks, learners, resamplings, paired = FALSE) {
   tasks = assert_tasks(as_tasks(tasks))
   learners = assert_learners(as_learners(learners))
   resamplings = assert_resamplings(as_resamplings(resamplings))
 
-  grid = CJ(task = seq_along(tasks), resampling = seq_along(resamplings))
-  is_instantiated = map_lgl(resamplings, "is_instantiated")
+  if (assert_flag(paired)) {
+    if (length(tasks) != length(resamplings)) {
+      stopf("If `paired` is `TRUE`, you need to provide the same number of tasks and instantiated resamplings")
+    }
 
-  if (any(is_instantiated)) {
-    task_nrow = unique(map_int(tasks, "nrow"))
-    if (length(task_nrow) != 1L) {
-      stopf("All resamplings must be uninstantiated, or must operate on tasks with the same number of rows")
+    for (i in seq_along(tasks)) {
+      task = tasks[[i]]
+      resampling = resamplings[[i]]
+      if (!resamplings[[i]]$is_instantiated) {
+        stopf("Resampling #%i ('%s' for task '%s') is not instantiated", i, resampling$id, task$id)
+      }
+      if (resampling$task_hash != task$hash) {
+        stopf("Resampling #%i ('%s' for task '%s') is not instantiated on the corresponding task", i, resampling$id, task$id)
+      }
     }
-    if (!identical(task_nrow, unique(map_int(resamplings, "task_nrow")))) {
-      stop("A Resampling is instantiated for a task with a different number of observations")
-    }
-    instances = pmap(grid, function(task, resampling) resamplings[[resampling]]$clone())
+
+    grid = CJ(task = seq_along(tasks), learner = seq_along(learners))
+    grid$instance = seq_row(grid)
+
+    tab = data.table(
+      task = tasks[grid$task],
+      learner = learners[grid$learner],
+      resampling = resamplings[grid$task]
+    )
   } else {
-    instances = pmap(grid, function(task, resampling) resamplings[[resampling]]$clone()$instantiate(tasks[[task]]))
+    grid = CJ(task = seq_along(tasks), resampling = seq_along(resamplings))
+    is_instantiated = map_lgl(resamplings, "is_instantiated")
+
+    if (any(is_instantiated)) {
+      task_nrow = unique(map_int(tasks, "nrow"))
+      if (length(task_nrow) != 1L) {
+        stopf("All resamplings must be uninstantiated, or must operate on tasks with the same number of rows")
+      }
+      if (!identical(task_nrow, unique(map_int(resamplings, "task_nrow")))) {
+        stop("A Resampling is instantiated for a task with a different number of observations")
+      }
+      instances = pmap(grid, function(task, resampling) resamplings[[resampling]]$clone())
+    } else {
+      instances = pmap(grid, function(task, resampling) resamplings[[resampling]]$clone()$instantiate(tasks[[task]]))
+    }
+
+    grid$instance = seq_row(grid)
+    grid = grid[CJ(task = seq_along(tasks), learner = seq_along(learners)), on = "task", allow.cartesian = TRUE]
+
+    tab = data.table(task = tasks[grid$task], learner = learners[grid$learner], resampling = instances[grid$instance])
   }
 
-  grid$instance = seq_row(grid)
-  grid = grid[CJ(task = seq_along(tasks), learner = seq_along(learners)), on = "task", allow.cartesian = TRUE]
-
-  tab = data.table(task = tasks[grid$task], learner = learners[grid$learner], resampling = instances[grid$instance])
   set_data_table_class(tab, "benchmark_grid")
-  tab
+  return(tab)
 }
 
 #' @export